backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
boto3==1.29.7
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
moto==4.2.14
httpx==0.25.2
python-multipart==0.0.6
tenacity==8.2.3
slowapi==0.1.9
# RAG dependencies
openai==1.3.0
numpy==1.24.3
scikit-learn==1.3.2
